+++ app/src/main/kotlin/studio/lunabee/compose/demo/presenter/pullToRefresh/PullToRefreshUiState.kt
@@ -21,7 +21,7 @@
 
 package studio.lunabee.compose.demo.presenter.pullToRefresh
 
-import studio.lunabee.compose.foundation.presenter.PresenterUiState
+import studio.lunabee.compose.presenter.PresenterUiState
 
 data class PullToRefreshUiState(
     val isRefreshing: Boolean,
