+++ lbcpresenter/lbcpresenter/src/main/kotlin/studio/lunabee/compose/presenter/LBSinglePresenter.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Lunabee Studio
+ * Copyright (c) 2025 Lunabee Studio
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +16,10 @@
  * LBSinglePresenter.kt
  * Lunabee Compose
  *
- * Created by Lunabee Studio / Date - 12/16/2024 - for the Lunabee Compose library.
+ * Created by Lunabee Studio / Date - 3/19/2025 - for the Lunabee Compose library.
  */
 
-package studio.lunabee.compose.foundation.presenter
+package studio.lunabee.compose.presenter
 
 /**
  * Subclass of [LBPresenter] to implement a single state presenter
