+++ lbcpresenter/lbcpresenter/src/main/kotlin/studio/lunabee/compose/presenter/PresenterUiState.kt
@@ -13,24 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * ReduceResult.kt
+ * PresenterUiState.kt
  * Lunabee Compose
  *
  * Created by Lunabee Studio / Date - 11/4/2024 - for the Lunabee Compose library.
  */
 
-package studio.lunabee.compose.foundation.presenter
+package studio.lunabee.compose.presenter
 
-data class ReduceResult<UiState>(
-    val uiState: UiState,
-    val sideEffect: ReduceSideEffect?,
-)
-
-infix fun <UiState : PresenterUiState> UiState.withSideEffect(sideEffect: ReduceSideEffect): ReduceResult<UiState> =
-    ReduceResult(this, sideEffect)
-
-fun <UiState : PresenterUiState> UiState.asResult(): ReduceResult<UiState> {
-    return ReduceResult(this, null)
-}
-
-typealias ReduceSideEffect = suspend () -> Unit
+interface PresenterUiState
